Multiple Fixes/Updates * More changes for Python3 compatibility * Move init script to be run by progexec thread, it is where it belongs * Change the default for UdpBroadcast to False, is not reliable with slow   clients, they drop packets * machif(s) now announce their name upon connection, removing UI need for   special knowledge * Fix bug with TinyG where a bad json string is sent form machif and slow   doesn't correctly count device queue * Got machif_smoothie running again, but have no HW to test it * Made sure to use pickle protocol 2, until we can move all code to python3 * Fix bug with jogging spindle button action * Optimize updating code in machine DRO wnd

+++ modules/wnd_jogging.py
@@ -658,12 +658,12 @@ class gsatJoggingPanel(wx.ScrolledWindow):
 
     def OnSpindleCWOn(self, e):
         speed = self.spindleSpeedSpinCtrl.GetValue()
-        speed_cmd = "{} S{:d}\n".format(gc.DEVICE_CMD_SPINDLE_CW_ON, round(speed))
+        speed_cmd = "{} S{:d}\n".format(gc.DEVICE_CMD_SPINDLE_CW_ON, int(round(speed)))
         self.mainWindow.SerialWrite(speed_cmd)
 
     def OnSpindleCCWOn(self, e):
         speed = self.spindleSpeedSpinCtrl.GetValue()
-        speed_cmd = "{} S{:d}\n".format(gc.DEVICE_CMD_SPINDLE_CCW_ON, round(speed))
+        speed_cmd = "{} S{:d}\n".format(gc.DEVICE_CMD_SPINDLE_CCW_ON, int(round(speed)))
         self.mainWindow.SerialWrite(speed_cmd)
 
     def OnSpindleOff(self, e):
